Added 2001-2011 village code mapping
All 2011-2001 village code maps were from http://egovstandards.gov.in/content/mapping-land-region-codes . Unfortunately this page no longer seems active. To match this data to the codes in the associated geojsons, you will need to concatenate state,district,tehsil and village codes for 2001 and use that as the 'join' field, so to speak, to match the two datasets.